Moved the update_model_lists_if_specified and update_model_lists methods into a special server hook, so they are executed only once even if gunicorn is started with more than one worker.
